requirements.txt
ezdxf==1.1.0
fonttools==4.42.1
numpy==1.26.0
pyparsing==3.1.1
pyserial==3.5
typing_extensions==4.8.0
